move to dashboard utils file
